requirements.txt
requests 
pymongo
python-dotenv
fastapi 
uvicorn # this is the ASGI server that FastAPI uses